src/redux/usersReducer.test.ts
import { UserType } from "../types/types";
import usersReducer, { actions, InitialState } from "./usersReducer";

let state: InitialState;

beforeEach(() => {
  state = {
    users: [
      {
        id: 0,
        name: "Roma0",
        followed: false,
        photos: { small: null, large: null },
        status: "status 0",
      },
      {
        id: 1,
        name: "Roma1",
        followed: false,
        photos: { small: null, large: null },
        status: "status 1",
      },
      {
        id: 2,
        name: "Roma2",
        followed: true,
        photos: { small: null, large: null },
        status: "status 2",
      },
      {
        id: 3,
        name: "Roma3",
        followed: true,
        photos: { small: null, large: null },
        status: "status 3",
      },
    ],
    pageSize: 5,
    totalUsersCount: 0,
    currentPage: 3,
    isFetching: false,
    followingProgress: [],
    portionOfPagesNumber: 1,
  };
});

test("follow success", () => {
  const newState = usersReducer(state, actions.followSucces(1));

  expect(newState.users[0].followed).toBeFalsy();
  expect(newState.users[1].followed).toBeTruthy();
});

test("unfollow success", () => {
  const newState = usersReducer(state, actions.unfollowSucces(3));

  expect(newState.users[2].followed).toBeTruthy();
  expect(newState.users[3].followed).toBeFalsy();
});
